refactor: Extract horizontal stacked bar chart into separate component
- Create HorizontalStackedBarAdhoc component following existing pattern
- Extract stacked bar logic from AdhocChart switch statement
- Improve code organization and reusability
- Maintain existing functionality and ref forwarding for export
